+++ CHANGES.md
@@ -1,7 +1,9 @@
-### Version dev-1.1.0-0
+### Version 1.1.0
 - [IMPORTANT] Introduce LicenseService to limit several functionalities for FREE version
 - [FEATURE] In Free version you can only create 3 configurations
 - [FEATURE] AutoClose Popup, if clicked outside
+- [FEATURE] Add new appearance behaviour: always
+- [FEATURE] Add new positions: center left, center, right, bottom left, bottom center and bottom right
 
 ### Version 1.0.3
 - [BUGFIX] ClosestConfigurationService->get need to be able to return no configuration, if not found